divided some code from TodoList into more components
